// src/commands/sites/sites-create.d.ts
import { OptionValues } from 'commander';
import BaseCommand from '../base-command.js';
export declare const getSiteNameInput: (name: any) => Promise<{
    name: any;
}>;
export declare const sitesCreate: (options: OptionValues, command: BaseCommand) => Promise<undefined>;
//# sourceMappingURL=sites-create.d.ts.map